# [OpenCL] Enable SVM for SGEMV Kernel