This patch updates FloatTensor to allocate SVM (Shared Virtual Memory) for unmanaged memory.
This change enables the use of SVM in the SGEMV (Single Precision General Matrix-Vector Multiplication) kernel.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

## Benchmark Result

#### 1. Linux-x86 (NVIDIA GeForce RTX 1650)
| |Time|
|------|---|
|Before|47 ms|
|After|19.15 ms|


#### 2. Linux-x86 (Intel(R) UHD Graphics 770)
| |Time|
|------|---|
|Before|89.7 ms|
|After|45.25 ms|

#### 3. Android (Qualcomm Adreno 740)
| |Time|
|------|---|
|Before|145.8 ms|
|After|108.85ms|

## Note

The following matrix sizes are used for the benchmark and profile.
- M: 1
- K: 3072
- N: 21180